Have VW's QT search behave like ASP's
This means that QT_ARBITRARY_MODULES will need to apply their own
prefix of Qt. It also means that if the user needs to add per module
library link flags, they'll have to do it themselves with the
PKG_ARBITRARY_QT_LIBS option.

The previous form of QT_ARBITRARY was not OSX safe since it always
applied -lQt<MODULE> for every module, even when the user specified a
framework to be linked againsted.