geoblend: Split into multiple archives
This is speed up perceived compile time and to reduce instantaneous
memory usage. The thread to compile geoblend uses about 1.6G by itself
in gcc4.2. Now that it has been split down, no single thread uses over
700 MB.

I'm using archives only because it is the only way in automake to
compile the same source many times with different CPP
flags. Geoblend_help.cc is the source for each archive, it just takes
a macro to define which version it is instantiating.

The downsides of this method is that more hard drive space is used (an
additional 40MB) and the Makefile.am looks a little ugly.